# api/grpc/hello_world_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "grpc/hello_world_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "helloWorldAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/get/{key}": {
      "get": {
        "operationId": "helloWorldAPI_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hello_world_apiValue"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "helloWorldAPI"
        ]
      }
    },
    "/v1/set": {
      "post": {
        "operationId": "helloWorldAPI_Set",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hello_world_apiSetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/hello_world_apiSetRequest"
            }
          }
        ],
        "tags": [
          "helloWorldAPI"
        ]
      }
    },
    "/v1/version": {
      "post": {
        "operationId": "helloWorldAPI_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hello_world_apiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/hello_world_apiVersionRequest"
            }
          }
        ],
        "tags": [
          "helloWorldAPI"
        ]
      }
    }
  },
  "definitions": {
    "hello_world_apiSetRequest": {
      "type": "object",
      "properties": {
        "vals": {
          "$ref": "#/definitions/hello_world_apiValues"
        }
      }
    },
    "hello_world_apiSetResponse": {
      "type": "object"
    },
    "hello_world_apiValue": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "hello_world_apiValues": {
      "type": "object",
      "properties": {
        "values": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/hello_world_apiValue"
          }
        }
      }
    },
    "hello_world_apiVersionRequest": {
      "type": "object"
    },
    "hello_world_apiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}